chore: rename Networth to PortfolioBalance for clarity
